style: 默认皮肤极简现代落地（mockup v3 选定风格）——DARK/LIGHT 主题重新 token：VS Code 风深色底+单一蓝强调(#3794ff/#4fc1ff)+细分割线(#3c3c3c)，状态栏/补全/选中/用户块全量扁平化，语义色沿用基底契约不破；FloatBox 边框 double→single（1px 细分割线）；全量零回归

+++ src/wxnodus-ui/components/floating.tsx
@@ -7,7 +7,7 @@ export function FloatBox({
   children,
   color,
   display,
-  borderStyle = 'double',
+  borderStyle = 'single',
   noBorder = false
 }: {
   children: ReactNode
